# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Care-Cordination server API documentation.",
        "title": "Care-Cordination API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "your-email@domain.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/location-transfers": {
            "get": {
                "description": "List all location transfers with pagination and search",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "LocationTransfer"
                ],
                "summary": "List location transfers",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size (default: 10, max: 100)",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search by client name",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/resp.SuccessResponse-resp_PaginationResponse-location_transfer_ListLocationTransfersResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Register a new location transfer for a client",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "LocationTransfer"
                ],
                "summary": "Register a location transfer",
                "parameters": [
                    {
                        "description": "Location Transfer Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/location_transfer.RegisterLocationTransferRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/resp.SuccessResponse-location_transfer_RegisterLocationTransferResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/locations": {
            "get": {
                "description": "List all locations with pagination and search",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Location"
                ],
                "summary": "List locations",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page size (default: 10, max: 100)",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search by name, postal code, or address",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/resp.SuccessResponse-resp_PaginationResponse-array_locations_ListLocationsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new location",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Location"
                ],
                "summary": "Create a location",
                "parameters": [
                    {
                        "description": "Location",
                        "name": "location",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/locations.CreateLocationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/resp.SuccessResponse-locations_CreateLocationResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/resp.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "location_transfer.ListLocationTransfersResponse": {
            "type": "object",
            "properties": {
                "clientFirstName": {
                    "type": "string"
                },
                "clientId": {
                    "type": "string"
                },
                "clientLastName": {
                    "type": "string"
                },
                "currentCoordinatorFirstName": {
                    "type": "string"
                },
                "currentCoordinatorId": {
                    "type": "string"
                },
                "currentCoordinatorLastName": {
                    "type": "string"
                },
                "fromLocationId": {
                    "type": "string"
                },
                "fromLocationName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "newCoordinatorFirstName": {
                    "type": "string"
                },
                "newCoordinatorId": {
                    "type": "string"
                },
                "newCoordinatorLastName": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "toLocationId": {
                    "type": "string"
                },
                "toLocationName": {
                    "type": "string"
                },
                "transferDate": {
                    "description": "or time.Time, but for JSON, string",
                    "type": "string"
                }
            }
        },
        "location_transfer.RegisterLocationTransferRequest": {
            "type": "object",
            "required": [
                "clientId",
                "newCoordinatorId",
                "newLocationId"
            ],
            "properties": {
                "clientId": {
                    "type": "string"
                },
                "newCoordinatorId": {
                    "type": "string"
                },
                "newLocationId": {
                    "type": "string"
                }
            }
        },
        "location_transfer.RegisterLocationTransferResponse": {
            "type": "object",
            "properties": {
                "transferId": {
                    "type": "string"
                }
            }
        },
        "locations.CreateLocationRequest": {
            "type": "object",
            "required": [
                "address",
                "name",
                "postalCode"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "capacity": {
                    "type": "integer",
                    "minimum": 1
                },
                "name": {
                    "type": "string"
                },
                "occupied": {
                    "type": "integer",
                    "minimum": 0
                },
                "postalCode": {
                    "type": "string"
                }
            }
        },
        "locations.CreateLocationResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "locations.ListLocationsResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "capacity": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "occupied": {
                    "type": "integer"
                },
                "postalCode": {
                    "type": "string"
                }
            }
        },
...
        "resp.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "error message"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "resp.PaginationResponse-array_locations_ListLocationsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "$ref": "#/definitions/locations.ListLocationsResponse"
                        }
                    }
                },
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "totalCount": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                }
            }
        },
...
        "resp.PaginationResponse-location_transfer_ListLocationTransfersResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/location_transfer.ListLocationTransfersResponse"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "totalCount": {
                    "type": "integer"
                },
                "totalPages": {
                    "type": "integer"
                }
            }
        },
...
        "resp.SuccessResponse-location_transfer_RegisterLocationTransferResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/location_transfer.RegisterLocationTransferResponse"
                },
                "message": {
                    "type": "string",
                    "example": "success message"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "resp.SuccessResponse-locations_CreateLocationResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/locations.CreateLocationResponse"
                },
                "message": {
                    "type": "string",
                    "example": "success message"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "resp.SuccessResponse-resp_PaginationResponse-array_locations_ListLocationsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/resp.PaginationResponse-array_locations_ListLocationsResponse"
                },
                "message": {
                    "type": "string",
                    "example": "success message"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "resp.SuccessResponse-resp_PaginationResponse-location_transfer_ListLocationTransfersResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/resp.PaginationResponse-location_transfer_ListLocationTransfersResponse"
                },
                "message": {
                    "type": "string",
                    "example": "success message"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Add 'Bearer ' prefix before your JWT token for authentication",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "security": [
        {
            "Bearer": []
        }
    ]
}